pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image2md"
version = "0.1.0"
description = "Convert images to markdown with various converters including OCR, Vision, Structure, Azure Document Intelligence, and LLM-based approaches"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Priscila Saboia Moreira", email = "pmoreira@nd.edu"}
]
keywords = ["OCR", "markdown", "Azure", "OpenAI", "image processing"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
    "openai>=1.0.0"
]

[project.optional-dependencies]
azure = [
    "azure-ai-documentintelligence>=1.0.0",
    "azure-core>=1.24.0"
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.51.0"
]
all = [
    "azure-ai-documentintelligence>=1.0.0",
    "azure-core>=1.24.0",
    "openai>=1.0.0",
    "anthropic>=0.51.0"
]

[project.scripts]
image2md = "image2md.cli:main"
